pyproject.toml
[project]
name = "mars-v2"
version = "1.0.0"
requires-python = ">=3.10"

dependencies = [
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "pandas>=2.1.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
